Cyrille: commit page détail catégorie et page mon compte client

+++ Tp_Ecommerce/src/main/java/fr/adaming/dao/ClientDAOImpl.java
@@ -33,7 +33,7 @@ public class ClientDAOImpl implements IClientDAO {
 	public Client modifyClient(Client c) {
 		Session s = sf.getCurrentSession();
 
-		Client clOut = (Client) s.get(Client.class, c.getId());
+		Client clOut = (Client) s.get(Client.class, (int) c.getId());
 		clOut.setNom(c.getNom());
 		clOut.setMail(c.getMail());
 		clOut.setAdresse(c.getAdresse());
